+++ util/database.js
@@ -1,5 +1,5 @@
 const {Sequelize}=require('sequelize');
-const sequelize=new Sequelize('expenseapp','root','Subhadeep@1990',{
-    dialect:'mysql',host:'localhost',dialectModule:require('mysql4')
+const sequelize=new Sequelize(process.env.DB_NAME,process.env.DB_USERNAME,process.env.DB_PASSWORD,{
+    dialect:'mysql',host:process.env.DB_HOST,dialectModule:require('mysql4')
 });
 module.exports=sequelize;
